requirements.txt
minio
Flask==1.0.2
requests
python-dotenv